Move subcommands to their own packages
Otherwise the directory will contain many files in the future,
especially once tests are added.

Having a single command per file still allows to quickly jump to a
command code, but they are now in sub-packages.
