+++ pkg/cmd/cluster/cluster.go
@@ -1,12 +1,12 @@
-package cmd
+package cluster
 
 import (
 	"github.com/dcos/dcos-cli/pkg/cli"
 	"github.com/spf13/cobra"
 )
 
-// newCmdCluster creates the `dcos cluster` subcommand.
-func newCmdCluster(ctx *cli.Context) *cobra.Command {
+// NewCommand creates the `dcos cluster` subcommand.
+func NewCommand(ctx *cli.Context) *cobra.Command {
 	cmd := &cobra.Command{
 		Use: "cluster",
 	}
