+++ pkg/cmd/cluster/cluster_attach.go
@@ -1,4 +1,4 @@
-package cmd
+package cluster
 
 import (
 	"github.com/dcos/dcos-cli/pkg/cli"
