+++ pkg/cmd/cluster/cluster_list.go
@@ -1,4 +1,4 @@
-package cmd
+package cluster
 
 import (
 	"encoding/json"
