+++ pkg/cmd/cluster/cluster_remove.go
@@ -1,4 +1,4 @@
-package cmd
+package cluster
 
 import (
 	"errors"
