+++ pkg/cmd/config/config_set.go
@@ -1,4 +1,4 @@
-package cmd
+package config
 
 import (
 	"github.com/dcos/dcos-cli/pkg/cli"
