+++ pkg/cmd/config/config_show.go
@@ -1,4 +1,4 @@
-package cmd
+package config
 
 import (
 	"fmt"
